refactor: changed location of cli functions
